extract shared ui components
