+++ packages/frontend/src/pages/interop/components/InteropTransferSizeCard.tsx
@@ -1,0 +1,26 @@
+import { PrimaryCard } from '~/components/primary-card/PrimaryCard'
+import type { TransferSizeDataPoint } from '~/server/features/scaling/interop/utils/getTransferSizeChartData'
+import { TransferSizeChart } from '../summary/components/charts/TransferSizeChart'
+import { Last24HoursBadge } from './Last24HoursBadge'
+
+export function InteropTransferSizeCard({
+  data,
+  isLoading,
+}: {
+  data: TransferSizeDataPoint[] | undefined
+  isLoading: boolean
+}) {
+  return (
+    <PrimaryCard className="flex h-[230px] flex-col" data-hide-overflow-x>
+      <div className="flex items-center gap-2">
+        <h2 className="font-bold text-heading-20">
+          Transfer size distribution
+        </h2>
+        <Last24HoursBadge />
+      </div>
+      <div className="min-h-0 flex-1">
+        <TransferSizeChart data={data ?? []} isLoading={isLoading} horizontal />
+      </div>
+    </PrimaryCard>
+  )
+}
