+++ packages/frontend/src/pages/interop/token-frameworks/components/FrameworkTransferSizeWidget.tsx
@@ -1,9 +1,7 @@
 import { useQuery } from '@tanstack/react-query'
-import { PrimaryCard } from '~/components/primary-card/PrimaryCard'
 import { useTRPC } from '~/trpc/React'
-import { TransferSizeChart } from '../../summary/components/charts/TransferSizeChart'
+import { InteropTransferSizeCard } from '../../components/InteropTransferSizeCard'
 import { useTokenFrameworksSelectedChains } from '../utils/TokenFrameworksSelectedChainsContext'
-import { Last24HoursBadge } from './Last24HoursBadge'
 
 export function FrameworkTransferSizeWidget() {
   const trpc = useTRPC()
@@ -16,20 +14,9 @@ export function FrameworkTransferSizeWidget() {
   )
 
   return (
-    <PrimaryCard className="flex h-[230px] flex-col" data-hide-overflow-x>
-      <div className="flex items-center gap-2">
-        <h2 className="font-bold text-heading-20">
-          Transfer size distribution
-        </h2>
-        <Last24HoursBadge />
-      </div>
-      <div className="min-h-0 flex-1">
-        <TransferSizeChart
-          data={data?.transferSizeChartData ?? []}
-          isLoading={isLoading}
-          horizontal
-        />
-      </div>
-    </PrimaryCard>
+    <InteropTransferSizeCard
+      data={data?.transferSizeChartData}
+      isLoading={isLoading}
+    />
   )
 }
