+++ packages/frontend/src/pages/interop/token-frameworks/components/TopTokensWidget.tsx
@@ -19,10 +19,10 @@ import type { TopTokenItem } from '~/server/features/scaling/interop/getTokenFra
 import { useTRPC } from '~/trpc/React'
 import { formatCurrency } from '~/utils/number-format/formatCurrency'
 import { formatInteger } from '~/utils/number-format/formatInteger'
+import { Last24HoursBadge } from '../../components/Last24HoursBadge'
 import { getInteropTokenUrl } from '../../utils/getInteropTokenUrl'
 import type { InteropTokenFramework } from '../getInteropTokenFrameworksData'
 import { useTokenFrameworksSelectedChains } from '../utils/TokenFrameworksSelectedChainsContext'
-import { Last24HoursBadge } from './Last24HoursBadge'
 import { TokenFrameworksTransferTrigger } from './TokenFrameworksTransferTrigger'
 
 export function TopTokensWidget({
